clean up

+++ os_stub/opt/utils/take_video.py
@@ -9,34 +9,36 @@ import picamera
 
 VIDEO_DURATION = 300 # s
 FPS = 2 # s
-RESOLUTION = (1280, 720) # s
+MIN_SPACE_BEFORE_DELETE_OLD = 10.0  # %
+SPACE_TO_LEAVE_AFTER_DELETE = 20.0  # %
+RESOLUTION = (1280, 972) # s
 VIDEO_DIR = "/videos/"
 PREVIEW = True
 
 logging.basicConfig()
 logging.getLogger().setLevel(logging.INFO)
 
-picam = picamera.PiCamera()
-picam.framerate = FPS
-picam.resolution = RESOLUTION
-time.sleep(2)
-
 
 def remove_old_files():
+
+    if available_disk_space() > MIN_SPACE_BEFORE_DELETE_OLD:
+        return
     all_files = []
     # should not be any , but remove them if the exist
     temp_files = []
-    for g in sorted(glob.glob(os.path.join(VIDEO_DIR, '*.h264'))):
-        if g.startswith("tmp.."):
+
+    for g in sorted(glob.glob(os.path.join(VIDEO_DIR, '*.h264')), reverse=True):
+        if g.startswith("tmp."):
             temp_files.append(g)
         elif g.endswith(".h264"):
             all_files.append(g)
-
-    # remove all but last temp files, in case
-    for r in temp_files[0: -1]:
+            # remove all but last temp files, in case
+    for r in temp_files[1: len(temp_files)]:
         os.remove(r)
-    # remove last img
-    os.remove(all_files[0])
+
+    while available_disk_space() < SPACE_TO_LEAVE_AFTER_DELETE:
+        # remove oldest images
+        os.remove(all_files.pop())
 
 
 def device_id():
@@ -51,6 +53,7 @@ def device_id():
         logging.warning("Could not read serial number setting to 00000000")
         cpuserial = "00000000"
     return cpuserial
+
 
 def available_disk_space():
     command = "df %s --output=used,size | tail -1 | tr -s ' '" % VIDEO_DIR
@@ -74,6 +77,13 @@ def make_video_name():
     return path, now
 
 
+remove_old_files()
+
+picam = picamera.PiCamera()
+picam.framerate = FPS
+picam.resolution = RESOLUTION
+time.sleep(2)
+
 video_path, video_start_time = make_video_name()
 logging.warning(f"Starting {video_path}")
 
@@ -87,12 +97,12 @@ picam.start_recording(video_path)
 
 while True:
     # if picam.recording:
-    if time.time() - video_start_time >  VIDEO_DURATION:
+    if time.time() - video_start_time > VIDEO_DURATION:
         new_video_path, new_video_time = make_video_name()
         picam.split_recording(new_video_path)
         logging.warning(f"Starting {new_video_path}")
-        if available_disk_space() < 10.0:
-            remove_old_files()
+        remove_old_files()
+
         video_path_final_name = os.path.join(VIDEO_DIR, os.path.basename(video_path)[4:])
         os.rename(video_path, video_path_final_name)
         logging.warning(f"Completed  {video_path_final_name}")
